+++ src/mechanics_dsl/__init__.py
@@ -39,7 +39,7 @@ from .symbolic import SymbolicEngine
 # Utils imports
 from .utils import config, logger, setup_logging
 
-__version__ = "2.1.1"
+__version__ = "2.1.2"
 __author__ = "Noah Parsons"
 __license__ = "MIT"